# docs/swagger.yaml
definitions:
  models.ErrorMessage:
    properties:
      error:
        type: string
    type: object
  models.Group:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    required:
    - name
    type: object
...
  models.Project:
    properties:
      created_by:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      language:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    required:
    - created_by
    - language
    - name
    type: object
  models.SuccessFindGroup:
    properties:
      data:
        $ref: '#/definitions/models.Project'
    type: object
  models.SuccessFindGroups:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Project'
        type: array
    type: object
...
  models.SuccessFindUser:
    properties:
      data:
        $ref: '#/definitions/models.User'
    type: object
  models.SuccessFindUsers:
    properties:
      data:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  models.SuccessGroupCreation:
    properties:
      data:
        $ref: '#/definitions/models.Project'
    type: object
  models.SuccessGroupDelete:
    properties:
      data:
        type: boolean
    type: object
  models.SuccessGroupUpdate:
    properties:
      data:
        $ref: '#/definitions/models.Project'
    type: object
...
  models.SuccessUserCreation:
    properties:
      data:
        $ref: '#/definitions/models.User'
    type: object
  models.SuccessUserDelete:
    properties:
      data:
        type: boolean
    type: object
  models.SuccessUserUpdate:
    properties:
      data:
        $ref: '#/definitions/models.User'
    type: object
  models.TokenRefresh:
    properties:
      token:
        type: string
    type: object
  models.TokenRequest:
    properties:
      mail:
        type: string
      password:
        type: string
    type: object
  models.TokenResponse:
    properties:
      token:
        type: string
    type: object
  models.User:
    properties:
      createdAt:
        type: string
      group:
        type: integer
      id:
        type: integer
      lastname:
        type: string
      mail:
        type: string
      name:
        type: string
      password:
        type: string
      state:
        type: boolean
      updatedAt:
        type: string
      username:
        type: string
    required:
    - group
    - lastname
    - mail
    - name
    - password
    - state
    - username
    type: object
info:
  contact: {}
paths:
  /groups:
    get:
      description: Bulk all nepackage's groups
      parameters:
      - description: JWT without bearer
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessFindGroups'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorMessage'
      summary: FindGroups
      tags:
      - Groups
    post:
      description: Create group with models.Group model
      parameters:
      - description: JWT without bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: Group data
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/models.Group'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessGroupCreation'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorMessage'
      summary: CreateGroup
      tags:
      - Groups
  /groups/{id}:
    delete:
      description: Delete group with using id
      parameters:
      - description: JWT without bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessGroupDelete'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorMessage'
      summary: DeleteGroup
      tags:
      - Groups
    get:
      description: Find a group with given id in path
      parameters:
      - description: JWT without bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessFindGroup'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorMessage'
      summary: FindGroup
      tags:
      - Groups
    patch:
      description: Update group with models.Group model
      parameters:
      - description: JWT without bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: Group data
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/models.Group'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessGroupUpdate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorMessage'
      summary: UpdateGroup
      tags:
      - Groups
...
  /token:
    post:
      consumes:
      - application/json
      description: Generate a valid token with user's mail and password
      parameters:
      - description: User credentials
        in: body
        name: auth
        required: true
        schema:
          $ref: '#/definitions/models.TokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TokenResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorMessage'
      summary: GenerateToken
      tags:
      - Authentication
  /token/refresh:
    post:
      consumes:
      - application/json
      description: Refresh JWT, validate that actual JWT is ok
      parameters:
      - description: User credentials
        in: body
        name: auth
        required: true
        schema:
          $ref: '#/definitions/models.TokenRefresh'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TokenRefresh'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorMessage'
      summary: RefreshToken
      tags:
      - Authentication
  /users:
    get:
      description: Bulk all nepackage's users
      parameters:
      - description: JWT without bearer
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessFindUsers'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorMessage'
      summary: FindUsers
      tags:
      - Users
    post:
      description: Create user with models.User model
      parameters:
      - description: JWT without bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessUserCreation'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorMessage'
      summary: CreateUser
      tags:
      - Users
  /users/{id}:
    delete:
      description: Delete user with using id
      parameters:
      - description: JWT without bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessUserDelete'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorMessage'
      summary: DeleteUser
      tags:
      - Users
    get:
      description: Find a user with given id in path
      parameters:
      - description: JWT without bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessFindUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorMessage'
      summary: FindUser
      tags:
      - Users
    patch:
      description: Update user with models.User model
      parameters:
      - description: JWT without bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessUserUpdate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorMessage'
      summary: UpdateUser
      tags:
      - Users
swagger: "2.0"